refactor(frontend): simplify session hydration test helpers
