+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions-stages-test-harness.ts
@@ -12,7 +12,6 @@ import {
   toAgentSessionPresenceSnapshotFromLiveSnapshot,
 } from "@openducktor/core";
 import type { SetStateAction } from "react";
-import { mergeHydratedMessages } from "@/state/operations/agent-orchestrator/support/hydrated-message-merge";
 import { getSessionMessageCount } from "@/state/operations/agent-orchestrator/support/messages";
 import { sessionMessageAt } from "@/test-utils/session-message-test-helpers";
 import type { AgentSessionState } from "@/types/agent-orchestrator";
@@ -28,6 +27,7 @@ import {
   type SessionLoadIntent,
   type UpdateSession,
 } from "./load-sessions-stages";
+import { createStateHarness } from "./load-sessions-state-test-harness";
 import { agentSessionPresenceLookupKey } from "./session-presence-cache";
 
 export type SessionStateMap = Record<string, AgentSessionState>;
@@ -142,37 +142,6 @@ export const createStalePresence = (externalSessionId: string, workingDirectory:
     snapshot: null,
   });
 
-export const createStateHarness = (sessions: Record<string, AgentSessionState>) => {
-  let state = sessions;
-  const sessionsRef = { current: state };
-  return {
-    sessionsRef,
-    setSessionsById: (
-      updater:
-        | Record<string, AgentSessionState>
-        | ((current: Record<string, AgentSessionState>) => Record<string, AgentSessionState>),
-    ) => {
-      state = typeof updater === "function" ? updater(state) : updater;
-      sessionsRef.current = state;
-    },
-    updateSession: (
-      externalSessionId: string,
-      updater: (current: AgentSessionState) => AgentSessionState,
-    ) => {
-      const current = state[externalSessionId];
-      if (!current) {
-        return;
-      }
-      state = {
-        ...state,
-        [externalSessionId]: updater(current),
-      };
-      sessionsRef.current = state;
-    },
-    getState: () => state,
-  };
-};
-
 export const createTaskFixture = (): TaskCard => ({
   id: "task-1",
   title: "Refactor loader",
@@ -245,9 +214,9 @@ export {
   agentSessionPresenceLookupKey,
   createHydrationPromptAssemblerStage,
   createRuntimeResolutionPlannerStage,
+  createStateHarness,
   getSessionMessageCount,
   hydrateSessionRecordsStage,
-  mergeHydratedMessages,
   preparePersistedSessionMergeStage,
   reconcileLiveSessionsStage,
   sessionMessageAt,
