+++ packages/frontend/src/state/operations/agent-orchestrator/lifecycle/load-sessions-test-harness.ts
@@ -87,37 +87,6 @@ export const getSession = (
   return session;
 };
 
-export const createStateHarness = (sessions: Record<string, AgentSessionState> = {}) => {
-  let state = sessions;
-  const sessionsRef = { current: state };
-  return {
-    sessionsRef,
-    setSessionsById: (
-      updater:
-        | Record<string, AgentSessionState>
-        | ((current: Record<string, AgentSessionState>) => Record<string, AgentSessionState>),
-    ) => {
-      state = typeof updater === "function" ? updater(state) : updater;
-      sessionsRef.current = state;
-    },
-    updateSession: (
-      externalSessionId: string,
-      updater: (current: AgentSessionState) => AgentSessionState,
-    ) => {
-      const current = state[externalSessionId];
-      if (!current) {
-        return;
-      }
-      state = {
-        ...state,
-        [externalSessionId]: updater(current),
-      };
-      sessionsRef.current = state;
-    },
-    getState: () => state,
-  };
-};
-
 export const persistedSessionRecord = (
   input: {
     externalSessionId: string;
@@ -303,6 +272,7 @@ export const waitForHistoryCallCount = async (
   await waitForHistoryCallCount(getHistoryCalls, expectedCalls, remainingAttempts - 1);
 };
 
+export { createStateHarness } from "./load-sessions-state-test-harness";
 export type {
   AgentSessionPresenceSnapshot,
   AgentSessionRecord,
